fix: friend classes devided by storages
