+++ src/main/java/ru/yandex/practicum/filmorate/storage/friendship/FriendshipStorage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.storage.friends;
+package ru.yandex.practicum.filmorate.storage.friendship;
 
 import ru.yandex.practicum.filmorate.model.Friendship;
